agregando whoops

+++ app/Services/JobService.php
@@ -1,11 +1,15 @@
 <?php
 namespace  app\Services;
 use app\Models\Job;
+use Aura\Router\Exception;
 
 class JobService{
   public function deleteJob($id)
   {
     $job = Job::where($id);
+    if(!$job){
+      throw new Exception("no exciste el Job");
+    }
     $job->delete();
   }
 }
